msm: kgsl: make per process stats atomic
There is a chance of messing up per process stats if something
like a free and allocation happen at the same time for memory
belonging to the same process. So make per process stats atomic.
Decrement process stats when mem entry is destroyed. Also make
sure that mapsize is updated for all the memdescs.

Change-Id: I5da1cf368d523768d0b267b314394612934a620d
Signed-off-by: Puranam V G Tejaswi <pvgtejas@codeaurora.org>